Closing #42 - a nasty SSL issue due different Java versions.
As described in ticket, sunlightfoundation.com has newer certificate,
not supported in Java 7, causing it to throw
SSLHandshakeException. Solution is to use http site, which is done on
Java 6/7; on Java 8 (or other versions, presuming newer ones) it will use
https site.

This commit also does a small refactoring of location-by-zip function.