Added support for setting database type in url and viewing pending migrations.
We are expecting and supporting two databases: mysql and h2, so setting
url will choose what database to use. Urls should be in form:

 * mysql://<host>:[optional <port>]/database
 * h2:file:///<absolute path>/database

H2 also supports `mem` and `tcp` as connection types (instead of
`file`), where for `tcp` user should use host with port (host:port)
instead of file name.

Viewing pending migrations is mainly for checking if there are
migrations that are expected to be run (detected new .sql or .java files
in migrations folder), but are not yet recorded in Flyway meta
table. With this, we can do some actions before migration is done,
without relying on Flyway callback system.