Updated DynamoDB sum and added NO_CHECKSUM option.
DynamoDB is relased in unversioned archive, and is hard to guess when it
gets updated. This commit updates checksum of new release, but also adds
checks for NO_CHECKSUM env variable: if set, sha1 sums will not be
checked against downloaded files.

This will make it easier to use, without explicitly updating script
every time DynamoDB gets a new release.